+++ R/plot.data_agro.R
@@ -5,8 +5,6 @@
 #' 
 #' @param x The data frame. It should come from \code{\link{format_data_PPBstats.data_agro}}
 #' 
-#' @param data_version data frame coming from \code{\link{format_data_PPBstats.data_agro_version}}
-#' 
 #' @param plot_type the type of plot you wish. It can be :
 #' \itemize{
 #'  \item "pam" for presence abscence matrix that represent the combinaison of germplasm x location
@@ -70,12 +68,8 @@
 #' 
 #' @export
 #' 
-#' @import ggplot2
-#' @import plyr
-#' 
 plot.data_agro = function(
   x,
-  data_version = NULL,
   plot_type = c("pam", "histogramm", "barplot", "boxplot", "interaction", "biplot", "radar", "raster", "map"),
   x_axis = NULL,
   in_col = NULL,
@@ -89,7 +83,6 @@ plot.data_agro = function(
 ){
   p_out = plot_descriptive_data(
     x,
-    data_version,
     plot_type,
     x_axis,
     in_col,