+++ R/plot.mean_comparisons_predict_the_past_model_bh_GxE.R
@@ -5,8 +5,6 @@
 #'
 #' @param x Output from \code{\link{mean_comparisons.predict_the_past_model_bh_GxE}} 
 #' 
-#' @param data_version Output from \code{\link{format_data_PPBstats.data_agro_version}} 
-#' 
 #' @param plot_type "interaction", "barplot" or "score"
 #' 
 #' @param nb_parameters_per_plot number of parameter per plot to display
@@ -28,7 +26,6 @@
 #' 
 plot.mean_comparisons_predict_the_past_model_bh_GxE <- function(
   x,
-  data_version = NULL,
   plot_type = "interaction",
   nb_parameters_per_plot = 8, ...
 ){
@@ -42,8 +39,6 @@ plot.mean_comparisons_predict_the_past_model_bh_GxE <- function(
   
   attributes(x)$PPBstats.object = "mean_comparisons_model_bh_intra_location"
   
-  if( !is.null(data_version) ) { stop("data_version must be NULL with data plot from predict_the_past()") }
-  
   if(plot_type == "score" | plot_type == "interaction") {
     ylab = paste(
       x$data_mean_comparisons[[1]]$mean.comparisons$entry,
@@ -54,7 +49,7 @@ plot.mean_comparisons_predict_the_past_model_bh_GxE <- function(
     x$data_mean_comparisons[[1]]$mean.comparisons$entry = ylab
   }
   
-  out = plot.mean_comparisons_model_bh_intra_location(x, data_version, plot_type, nb_parameters_per_plot)
+  out = plot.mean_comparisons_model_bh_intra_location(x, plot_type, nb_parameters_per_plot)
   
   if(plot_type == "barplot") {
     p1 = lapply(out$data_mean_comparisons[[1]], function(x){ x + geom_bar(aes(fill = parameter_statuts), stat = "identity") } )